fix: 타이머 관련 소켓 작동하도록 수정
