+++ client/src/components/modals/timerSetting/timerSetting.tsx
@@ -1,7 +1,5 @@
 import { useEffect, useRef } from "react";
 import { useParams } from "next/navigation";
-import { useSetAtom } from "jotai";
-import { timerAtom, isTimerVisibleAtom } from "@/jotai/atom";
 import { BaseModal } from "@/components/modal/baseModal/baseModal";
 import { chatSocket } from "@/socket";
 import { numberToString } from "@/utils";
@@ -14,8 +12,6 @@ interface IModal {
 
 const TimerSetting = ({ onClose }: IModal) => {
     const { roomId } = useParams();
-    const setTimer = useSetAtom(timerAtom);
-    const setIsTimerVisible = useSetAtom(isTimerVisibleAtom);
 
     const minRef = useRef<HTMLInputElement>({} as HTMLInputElement);
     const secRef = useRef<HTMLInputElement>({} as HTMLInputElement);
@@ -27,15 +23,14 @@ const TimerSetting = ({ onClose }: IModal) => {
 
     const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
         e.preventDefault();
-
         if (minRef.current.value === "00" && secRef.current.value === "00") return;
 
         const minute = Number(minRef.current.value) || 0;
         const second = Number(secRef.current.value) || 0;
 
-        setTimer({ minute, second });
-
         chatSocket.emit("start_timer", { roomId, duration: { minute, second } });
+
+        onClose();
     };
 
     const handleInput = (
@@ -56,19 +51,6 @@ const TimerSetting = ({ onClose }: IModal) => {
     ) => {
         targetRef.current!.value = e.target.value.padStart(2, "0");
     };
-
-    const handleStartTimer = () => {
-        onClose();
-        setIsTimerVisible(true);
-    };
-
-    useEffect(() => {
-        chatSocket.on("start_timer", handleStartTimer);
-
-        return () => {
-            chatSocket.off("start_timer");
-        };
-    }, [handleStartTimer]);
 
     return (
         <BaseModal onClose={onClose}>
